Move Pushtype to utils
